requirements.txt
flask
flask-cors
stripe
pymongo
werkzeug
psycopg2-binary